Define shared assembly dependency resolver boundary
Add a Roslyn-free assembly identity resolver callback in Metadata and a Services-owned rich dependency resolver with package, framework, sibling, deps.json, project, and corpus provenance.

Refactor DecompilerHarness and ReturnToSender to consume the shared resolver while keeping target exclusion and Roslyn MetadataReference creation local to the harness.
