+++ tools/DecompilerHarness/FidelityCheck.cs
@@ -5,8 +5,8 @@ using System.Reflection.Metadata;
 using System.Reflection.PortableExecutable;
 using System.Text;
 using System.Text.Json;
-using System.Xml.Linq;
 
+using DotnetInspector.Services;
 using ILInspector.Decompiler;
 using ILInspector.Decompiler.Pipeline;
 using ILInspector.Metadata;
@@ -3403,7 +3403,6 @@ static class FidelityCheck
     {
         var seen = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
         var referencePaths = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-        var targetName = Path.GetFileNameWithoutExtension(targetPath);
         var builder = ImmutableArray.CreateBuilder<MetadataReference>();
 
         void Add(string path)
@@ -3413,8 +3412,6 @@ static class FidelityCheck
             if (!File.Exists(path))
                 return;
             string simple = Path.GetFileNameWithoutExtension(path);
-            if (simple.Equals(targetName, StringComparison.OrdinalIgnoreCase))
-                return; // the target is reconstructed, not referenced
             if (seen.Contains(simple))
                 return; // first definition wins (prefer TPA over a dir copy)
             try
@@ -3433,577 +3430,15 @@ static class FidelityCheck
             catch (ArgumentException) { }
         }
 
-        foreach (var path in PackageDependencyReferencePaths(targetPath))
-            Add(path);
-
-        foreach (var path in (AppContext.GetData("TRUSTED_PLATFORM_ASSEMBLIES") as string ?? "")
-            .Split(Path.PathSeparator, StringSplitOptions.RemoveEmptyEntries))
-            Add(path);
-        AddSharedFrameworkReferences("Microsoft.AspNetCore.App", Add);
-
-        var dir = Path.GetDirectoryName(Path.GetFullPath(targetPath));
-        if (dir is not null && Directory.Exists(dir))
+        var resolver = new AssemblyDependencyResolver(new AssemblyDependencyResolutionOptions(targetPath)
         {
-            foreach (var path in Directory.EnumerateFiles(dir, "*.dll"))
-                Add(path);
-
-            AddDepsJsonReferences(dir, targetName, Add);
-        }
-
-        // Cross-package corpus siblings (the other assemblies under analysis) are
-        // not co-located with the target or on the TPA, so a body or sibling
-        // signature referencing them would not bind. Referencing them resolves
-        // those public types and lets SignatureAccessibility skip stubs exposing
-        // their internals (#1376).
-        if (corpusAssemblies is not null)
-            foreach (var path in corpusAssemblies)
-                Add(path);
+            CorpusAssemblyPaths = corpusAssemblies,
+            ExcludeTargetAssembly = true,
+        });
+        foreach (var dependency in resolver.ResolveAll())
+            Add(dependency.Path);
 
         return new ReferenceSet(builder.ToImmutable(), new SignatureAccessibility(referencePaths));
     }
 
-    static void AddSharedFrameworkReferences(string frameworkName, Action<string> add)
-    {
-        string? runtimeDir = Path.GetDirectoryName(typeof(object).Assembly.Location);
-        if (runtimeDir is null)
-            return;
-        string? sharedRoot = Directory.GetParent(runtimeDir)?.Parent?.FullName;
-        if (sharedRoot is null)
-            return;
-        string frameworkRoot = Path.Combine(sharedRoot, frameworkName);
-        if (!Directory.Exists(frameworkRoot))
-            return;
-
-        if (SelectSharedFrameworkDirectory(frameworkRoot, Path.GetFileName(runtimeDir)) is not { } selected)
-            return;
-        foreach (var path in Directory.EnumerateFiles(selected, "*.dll"))
-            add(path);
-    }
-
-    internal static string? SelectSharedFrameworkDirectory(string frameworkRoot, string runtimeVersion)
-    {
-        string exactDirectory = Path.Combine(frameworkRoot, runtimeVersion);
-        if (Directory.Exists(exactDirectory))
-            return exactDirectory;
-        if (!Version.TryParse(VersionCore(runtimeVersion), out var runtime))
-            return null;
-
-        return Directory.EnumerateDirectories(frameworkRoot)
-            .Select(directory => new
-            {
-                Directory = directory,
-                Version = Version.TryParse(VersionCore(Path.GetFileName(directory)), out var version) ? version : null,
-            })
-            .Where(candidate => candidate.Version is not null
-                && candidate.Version.Major == runtime.Major
-                && candidate.Version.Minor == runtime.Minor)
-            .OrderByDescending(candidate => candidate.Version)
-            .Select(candidate => candidate.Directory)
-            .FirstOrDefault();
-    }
-
-    static string VersionCore(string version) => version.Split('-', 2)[0];
-
-    internal static IReadOnlyList<string> PackageDependencyReferencePaths(string targetPath)
-        => PackageDependencyReferencePaths(targetPath, packageRoots: null);
-
-    internal static IReadOnlyList<string> PackageDependencyReferencePaths(string targetPath, IReadOnlyList<string>? packageRoots)
-    {
-        if (NuGetPackageContext(targetPath, packageRoots) is not { } context)
-            return [];
-
-        var nuspec = Directory.EnumerateFiles(context.PackageDirectory, "*.nuspec").FirstOrDefault();
-        if (nuspec is null)
-            return [];
-
-        var packageDirectories = new Dictionary<string, Dictionary<string, string>>(StringComparer.OrdinalIgnoreCase);
-        var dependencyGraph = new Dictionary<string, List<PackageDependency>>(StringComparer.OrdinalIgnoreCase);
-        List<PackageDependency> rootDependencies;
-        try
-        {
-            rootDependencies = CollectPackageDependencies(
-                context.PackageDirectory,
-                context.TargetFramework,
-                packageRoots,
-                packageDirectories,
-                dependencyGraph,
-                new HashSet<string>(StringComparer.OrdinalIgnoreCase));
-        }
-        catch (Exception ex) when (ex is IOException or UnauthorizedAccessException or System.Xml.XmlException)
-        {
-            return [];
-        }
-
-        var selectedVersions = packageDirectories.ToDictionary(
-            kv => kv.Key,
-            kv => kv.Value.Keys.OrderByDescending(version => version, PackageVersionComparer.Instance).First(),
-            StringComparer.OrdinalIgnoreCase);
-        var resolved = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (var dependency in rootDependencies)
-            AddResolvedPackage(dependency.Id);
-
-        var references = new List<string>();
-        foreach (string id in resolved.Order(StringComparer.OrdinalIgnoreCase))
-            foreach (var path in ProbeNuGetPackageVersionDlls(packageDirectories[id][selectedVersions[id]], context.TargetFramework))
-                references.Add(path);
-        return references;
-
-        void AddResolvedPackage(string id)
-        {
-            if (!selectedVersions.TryGetValue(id, out string? version) || !resolved.Add(id))
-                return;
-            if (!dependencyGraph.TryGetValue(PackageKey(id, version), out var dependencies))
-                return;
-            foreach (var dependency in dependencies)
-                AddResolvedPackage(dependency.Id);
-        }
-    }
-
-    static List<PackageDependency> CollectPackageDependencies(
-        string packageDirectory,
-        string targetFramework,
-        IReadOnlyList<string>? packageRoots,
-        Dictionary<string, Dictionary<string, string>> packageDirectories,
-        Dictionary<string, List<PackageDependency>> dependencyGraph,
-        HashSet<string> visiting)
-    {
-        var nuspec = Directory.EnumerateFiles(packageDirectory, "*.nuspec").FirstOrDefault();
-        if (nuspec is null)
-            return [];
-
-        var document = XDocument.Load(nuspec);
-        var dependencies = new List<PackageDependency>();
-        foreach (var dependency in SelectNuGetDependencies(document, targetFramework))
-        {
-            string? id = dependency.Attribute("id")?.Value;
-            string? version = DependencyExactVersion(dependency.Attribute("version")?.Value);
-            if (string.IsNullOrWhiteSpace(id) || string.IsNullOrWhiteSpace(version))
-                continue;
-
-            foreach (var root in NuGetPackageRoots(packageRoots))
-            {
-                var dependencyDirectory = Path.Combine(root, id.ToLowerInvariant(), version.ToLowerInvariant());
-                if (!Directory.Exists(dependencyDirectory))
-                    continue;
-
-                if (!packageDirectories.TryGetValue(id, out var versions))
-                    packageDirectories[id] = versions = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-                versions[version] = dependencyDirectory;
-                dependencies.Add(new PackageDependency(id, version));
-
-                string visitKey = $"{id}/{version}";
-                if (!dependencyGraph.ContainsKey(PackageKey(id, version)) && visiting.Add(visitKey))
-                {
-                    dependencyGraph[PackageKey(id, version)] = CollectPackageDependencies(
-                        dependencyDirectory,
-                        targetFramework,
-                        packageRoots,
-                        packageDirectories,
-                        dependencyGraph,
-                        visiting);
-                    visiting.Remove(visitKey);
-                }
-                break;
-            }
-        }
-        return dependencies;
-    }
-
-    sealed class PackageVersionComparer : IComparer<string>
-    {
-        public static readonly PackageVersionComparer Instance = new();
-
-        public int Compare(string? left, string? right)
-            => string.Equals(left, right, StringComparison.OrdinalIgnoreCase)
-                ? 0
-                : ComparePackageVersion(left ?? "", right ?? "");
-    }
-
-    readonly record struct PackageDependency(string Id, string Version);
-
-    static string PackageKey(string id, string version) => $"{id}/{version}";
-
-    static int ComparePackageVersion(string left, string right)
-    {
-        var leftVersion = PackageVersion.Parse(left);
-        var rightVersion = PackageVersion.Parse(right);
-
-        int releaseLength = Math.Max(leftVersion.Release.Length, rightVersion.Release.Length);
-        for (int i = 0; i < releaseLength; i++)
-        {
-            int leftPart = i < leftVersion.Release.Length ? leftVersion.Release[i] : 0;
-            int rightPart = i < rightVersion.Release.Length ? rightVersion.Release[i] : 0;
-            int comparison = leftPart.CompareTo(rightPart);
-            if (comparison != 0)
-                return comparison;
-        }
-
-        if (leftVersion.Prerelease.Length == 0 || rightVersion.Prerelease.Length == 0)
-            return leftVersion.Prerelease.Length == rightVersion.Prerelease.Length
-                ? 0
-                : leftVersion.Prerelease.Length == 0 ? 1 : -1;
-
-        int prereleaseLength = Math.Max(leftVersion.Prerelease.Length, rightVersion.Prerelease.Length);
-        for (int i = 0; i < prereleaseLength; i++)
-        {
-            if (i >= leftVersion.Prerelease.Length)
-                return -1;
-            if (i >= rightVersion.Prerelease.Length)
-                return 1;
-
-            var leftIdentifier = leftVersion.Prerelease[i];
-            var rightIdentifier = rightVersion.Prerelease[i];
-            int comparison = leftIdentifier.CompareTo(rightIdentifier);
-            if (comparison != 0)
-                return comparison;
-        }
-
-        return 0;
-    }
-
-    readonly record struct PackageVersion(ImmutableArray<int> Release, ImmutableArray<PrereleaseIdentifier> Prerelease)
-    {
-        public static PackageVersion Parse(string version)
-        {
-            string withoutBuild = version.Split('+', 2)[0];
-            var split = withoutBuild.Split('-', 2);
-            var release = split[0]
-                .Split('.', StringSplitOptions.RemoveEmptyEntries)
-                .Select(part => int.TryParse(part, out int value) ? value : 0)
-                .ToImmutableArray();
-            var prerelease = split.Length == 1
-                ? ImmutableArray<PrereleaseIdentifier>.Empty
-                : split[1]
-                    .Split('.', StringSplitOptions.RemoveEmptyEntries)
-                    .Select(PrereleaseIdentifier.Parse)
-                    .ToImmutableArray();
-            return new(release, prerelease);
-        }
-    }
-
-    readonly record struct PrereleaseIdentifier(bool IsNumeric, int NumericValue, string Text)
-    {
-        public static PrereleaseIdentifier Parse(string value)
-            => int.TryParse(value, out int numeric)
-                ? new(true, numeric, "")
-                : new(false, 0, value);
-
-        public int CompareTo(PrereleaseIdentifier other)
-        {
-            if (IsNumeric && other.IsNumeric)
-                return NumericValue.CompareTo(other.NumericValue);
-            if (IsNumeric != other.IsNumeric)
-                return IsNumeric ? -1 : 1;
-            return string.Compare(Text, other.Text, StringComparison.OrdinalIgnoreCase);
-        }
-    }
-
-    sealed record NuGetReferenceContext(
-        string PackageDirectory,
-        string TargetFramework);
-
-    static NuGetReferenceContext? NuGetPackageContext(string targetPath, IReadOnlyList<string>? packageRoots = null)
-    {
-        string fullPath = Path.GetFullPath(targetPath);
-        foreach (var root in NuGetPackageRoots(packageRoots))
-        {
-            string fullRoot = Path.GetFullPath(root).TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
-            string prefix = fullRoot + Path.DirectorySeparatorChar;
-            if (!fullPath.StartsWith(prefix, StringComparison.OrdinalIgnoreCase))
-                continue;
-
-            var parts = fullPath[prefix.Length..].Split(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar);
-            if (parts.Length < 5 || !parts[2].Equals("lib", StringComparison.OrdinalIgnoreCase))
-                return null;
-
-            return new(
-                Path.Combine(fullRoot, parts[0], parts[1]),
-                parts[3]);
-        }
-
-        return null;
-    }
-
-    static IEnumerable<string> NuGetPackageRoots(IReadOnlyList<string>? packageRoots = null)
-    {
-        if (packageRoots is not null)
-        {
-            foreach (var root in packageRoots)
-                yield return root;
-            yield break;
-        }
-
-        if (Environment.GetEnvironmentVariable("NUGET_PACKAGES") is { Length: > 0 } envRoot)
-            yield return envRoot;
-
-        string home = Environment.GetFolderPath(Environment.SpecialFolder.UserProfile);
-        if (!string.IsNullOrEmpty(home))
-            yield return Path.Combine(home, ".nuget", "packages");
-    }
-
-    static IEnumerable<XElement> SelectNuGetDependencies(XDocument document, string? tfm)
-    {
-        var dependencies = document.Descendants().FirstOrDefault(e => e.Name.LocalName == "dependencies");
-        if (dependencies is null)
-            yield break;
-
-        var directDependencies = dependencies.Elements().Where(e => e.Name.LocalName == "dependency").ToArray();
-        var groups = dependencies.Elements().Where(e => e.Name.LocalName == "group").ToArray();
-        string? normalizedTfm = NormalizeNuGetFramework(tfm);
-        var selectedGroup = normalizedTfm is null
-            ? null
-            : groups
-                .Select(group => new
-                {
-                    Group = group,
-                    Score = NuGetFrameworkCompatibilityScore(normalizedTfm, NormalizeNuGetFramework(group.Attribute("targetFramework")?.Value)),
-                })
-                .Where(candidate => candidate.Score is not null)
-                .OrderByDescending(candidate => candidate.Score)
-                .Select(candidate => candidate.Group)
-                .FirstOrDefault();
-
-        if (selectedGroup is not null)
-        {
-            foreach (var dependency in selectedGroup.Elements().Where(e => e.Name.LocalName == "dependency"))
-                yield return dependency;
-            yield break;
-        }
-
-        foreach (var dependency in directDependencies)
-            yield return dependency;
-    }
-
-    static string? NormalizeNuGetFramework(string? tfm)
-    {
-        if (string.IsNullOrWhiteSpace(tfm))
-            return null;
-
-        tfm = tfm.Trim().ToLowerInvariant();
-        if (tfm.StartsWith(".netstandard", StringComparison.Ordinal))
-            return "netstandard" + tfm[".netstandard".Length..];
-        if (tfm.StartsWith(".netcoreapp", StringComparison.Ordinal))
-            return "netcoreapp" + tfm[".netcoreapp".Length..];
-        if (tfm.StartsWith(".netframework", StringComparison.Ordinal))
-            return "net" + tfm[".netframework".Length..].Replace(".", "", StringComparison.Ordinal);
-        return tfm;
-    }
-
-    static int? NuGetFrameworkCompatibilityScore(string target, string? candidate)
-    {
-        if (candidate is null)
-            return null;
-        if (string.Equals(target, candidate, StringComparison.OrdinalIgnoreCase))
-            return 100_000;
-        if (!TryParseNuGetFramework(target, out var targetFramework) ||
-            !TryParseNuGetFramework(candidate, out var candidateFramework))
-            return null;
-
-        int versionScore = candidateFramework.Major * 100 + candidateFramework.Minor;
-        return targetFramework.Family switch
-        {
-            "net" when candidateFramework.Family == "net" && candidateFramework.VersionScore <= targetFramework.VersionScore
-                => 90_000 + versionScore,
-            "net" when candidateFramework.Family == "netcoreapp" && candidateFramework.VersionScore <= targetFramework.VersionScore
-                => 80_000 + versionScore,
-            "net" when candidateFramework.Family == "netstandard" && candidateFramework.VersionScore <= 201
-                => 70_000 + versionScore,
-            "netcoreapp" when candidateFramework.Family == "netcoreapp" && candidateFramework.VersionScore <= targetFramework.VersionScore
-                => 90_000 + versionScore,
-            "netcoreapp" when candidateFramework.Family == "netstandard" && candidateFramework.VersionScore <= 201
-                => 80_000 + versionScore,
-            "netstandard" when candidateFramework.Family == "netstandard" && candidateFramework.VersionScore <= targetFramework.VersionScore
-                => 90_000 + versionScore,
-            _ => null,
-        };
-    }
-
-    readonly record struct NuGetFramework(string Family, int Major, int Minor)
-    {
-        public int VersionScore => Major * 100 + Minor;
-    }
-
-    static bool TryParseNuGetFramework(string tfm, out NuGetFramework framework)
-    {
-        framework = default;
-        if (tfm.StartsWith("netstandard", StringComparison.Ordinal))
-            return TryParseVersionedFramework("netstandard", tfm["netstandard".Length..], out framework);
-        if (tfm.StartsWith("netcoreapp", StringComparison.Ordinal))
-            return TryParseVersionedFramework("netcoreapp", tfm["netcoreapp".Length..], out framework);
-        if (tfm.StartsWith("net", StringComparison.Ordinal) && tfm[3..].Contains('.', StringComparison.Ordinal))
-            return TryParseVersionedFramework("net", tfm["net".Length..], out framework);
-        return false;
-    }
-
-    static bool TryParseVersionedFramework(string family, string version, out NuGetFramework framework)
-    {
-        framework = default;
-        var parts = version.Split('.', 3);
-        if (parts.Length == 0 || !int.TryParse(parts[0], out int major))
-            return false;
-        int minor = 0;
-        if (parts.Length >= 2 && !int.TryParse(parts[1], out minor))
-            return false;
-        framework = new(family, major, minor);
-        return true;
-    }
-
-    static string? DependencyExactVersion(string? version)
-    {
-        if (string.IsNullOrWhiteSpace(version))
-            return null;
-
-        version = version.Trim();
-        if (version.Length >= 5 && version[0] == '[' && version[^1] == ']')
-        {
-            string inner = version[1..^1].Trim();
-            var range = inner.Split(',', 2);
-            if (range.Length == 1)
-                return inner;
-            if (range.Length == 2
-                && string.Equals(range[0].Trim(), range[1].Trim(), StringComparison.OrdinalIgnoreCase))
-                return range[0].Trim();
-        }
-
-        return version.IndexOfAny(['[', ']', '(', ')', ',']) < 0 ? version : null;
-    }
-
-    static IEnumerable<string> ProbeNuGetPackageVersionDlls(string packageDir, string? tfm)
-    {
-        if (!Directory.Exists(packageDir))
-            yield break;
-
-        foreach (string assetKind in (string[])["ref", "lib"])
-        {
-            if (tfm is not null && AssetDirectory(packageDir, assetKind, tfm) is { } exactAssetDir)
-            {
-                foreach (var path in Directory.EnumerateFiles(exactAssetDir, "*.dll"))
-                    yield return path;
-                yield break;
-            }
-        }
-
-        foreach (string assetKind in (string[])["ref", "lib"])
-        {
-            if (tfm is not null && CompatibleAssetDirectory(packageDir, assetKind, tfm) is { } compatibleAssetDir)
-            {
-                foreach (var path in Directory.EnumerateFiles(compatibleAssetDir, "*.dll"))
-                    yield return path;
-                yield break;
-            }
-        }
-    }
-
-    static string? AssetDirectory(string packageDir, string assetKind, string tfm)
-    {
-        string assetDir = Path.Combine(packageDir, assetKind, tfm);
-        return Directory.Exists(assetDir) ? assetDir : null;
-    }
-
-    static string? CompatibleAssetDirectory(string packageDir, string assetKind, string targetTfm)
-    {
-        string assetRoot = Path.Combine(packageDir, assetKind);
-        if (!Directory.Exists(assetRoot))
-            return null;
-
-        string? normalizedTarget = NormalizeNuGetFramework(targetTfm);
-        if (normalizedTarget is null)
-            return null;
-
-        return Directory.EnumerateDirectories(assetRoot)
-            .Select(dir => new
-            {
-                Directory = dir,
-                Score = NuGetFrameworkCompatibilityScore(normalizedTarget, NormalizeNuGetFramework(Path.GetFileName(dir))),
-            })
-            .Where(candidate => candidate.Score is not null)
-            .OrderByDescending(candidate => candidate.Score)
-            .Select(candidate => candidate.Directory)
-            .FirstOrDefault();
-    }
-
-    static void AddDepsJsonReferences(string targetDirectory, string targetName, Action<string> addReference)
-    {
-        var depsPath = Path.Combine(targetDirectory, $"{targetName}.deps.json");
-        if (!File.Exists(depsPath))
-            return;
-
-        try
-        {
-            using var doc = JsonDocument.Parse(File.ReadAllText(depsPath));
-            var root = doc.RootElement;
-            if (!root.TryGetProperty("targets", out var targets) ||
-                targets.ValueKind != JsonValueKind.Object ||
-                !root.TryGetProperty("libraries", out var libraries) ||
-                libraries.ValueKind != JsonValueKind.Object)
-                return;
-
-            var libraryPaths = new Dictionary<string, string>(StringComparer.OrdinalIgnoreCase);
-            foreach (var library in libraries.EnumerateObject())
-            {
-                if (library.Value.ValueKind == JsonValueKind.Object &&
-                    library.Value.TryGetProperty("path", out var pathElement) &&
-                    pathElement.ValueKind == JsonValueKind.String &&
-                    pathElement.GetString() is { Length: > 0 } path)
-                    libraryPaths[library.Name] = path;
-            }
-
-            foreach (var target in targets.EnumerateObject())
-            {
-                if (target.Value.ValueKind != JsonValueKind.Object)
-                    continue;
-
-                foreach (var library in target.Value.EnumerateObject())
-                {
-                    AddAssetGroup(targetDirectory, libraryPaths, library, "compile", addReference);
-                    AddAssetGroup(targetDirectory, libraryPaths, library, "runtime", addReference);
-                }
-            }
-        }
-        catch (IOException) { }
-        catch (UnauthorizedAccessException) { }
-        catch (JsonException) { }
-    }
-
-    static void AddAssetGroup(
-        string targetDirectory,
-        IReadOnlyDictionary<string, string> libraryPaths,
-        JsonProperty library,
-        string groupName,
-        Action<string> addReference)
-    {
-        if (!library.Value.TryGetProperty(groupName, out var assets))
-            return;
-        if (assets.ValueKind != JsonValueKind.Object)
-            return;
-
-        foreach (var asset in assets.EnumerateObject())
-        {
-            if (asset.Name == "_._")
-                continue;
-
-            if (asset.Value.ValueKind == JsonValueKind.Object &&
-                asset.Value.TryGetProperty("localPath", out var localPathElement) &&
-                localPathElement.ValueKind == JsonValueKind.String &&
-                localPathElement.GetString() is { Length: > 0 } localPath)
-                addReference(Path.Combine(targetDirectory, NativePath(localPath)));
-
-            if (libraryPaths.TryGetValue(library.Name, out var packagePath))
-                addReference(Path.Combine(GlobalPackagesRoot(), NativePath(packagePath), NativePath(asset.Name)));
-        }
-    }
-
-    static string NativePath(string path) => path.Replace('/', Path.DirectorySeparatorChar);
-
-    static string GlobalPackagesRoot()
-    {
-        var packagesRoot = Environment.GetEnvironmentVariable("NUGET_PACKAGES");
-        if (!string.IsNullOrEmpty(packagesRoot))
-            return packagesRoot;
-
-        return Path.Combine(
-            Environment.GetFolderPath(Environment.SpecialFolder.UserProfile),
-            ".nuget",
-            "packages");
-    }
 }
